reimplemented statistics without jtablesaw

+++ core/src/main/java/be/bavodaniels/finance/accounting/Accounting.java
@@ -5,7 +5,8 @@ import be.bavodaniels.finance.strategy.Statistics;
 import java.time.LocalDate;
 
 public interface Accounting {
-    void register(LocalDate date, Double price, Double underlyingPrice, int contractsHeld);
+
+    void register(LocalDate date, Double price, Double underlyingPrice, Double contractsHeld);
 
     Statistics getStatistics(double allocatedCapital);
 }
